fix : class adjusts

+++ src/main/java/com/lum/projetoTJW/models/Professor.java
@@ -5,19 +5,20 @@ import java.util.List;
 @Entity
 @Table(name = "tb_professor")
 public class Professor {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String nome;
     private String email;
 
     public Professor() {
     }
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -36,6 +37,9 @@ public class Professor {
     public void setEmail(String email) {
         this.email = email;
     }
+
+
+
 
 
 }
